Updated: tried to extend the memory space for arrays by defining the size type from int to unsigned long long
git-svn-id: svn+ssh://svn.gersteinlab.org/repos/bioinfo/libbios@666 2fd4de08-f6cc-4d2b-b124-baeea9ba92f3

+++ lib/bios/array.c
@@ -76,7 +76,7 @@ Array uArrayCreate(int n, int size)
 static void arrayExtend (Array a, int n)
 {
 	char *new ;
-	int oldsize, newsize ;
+	unsigned long long oldsize, newsize ;
 
 	if (!a || n < a->dim)
 		return ;
